lib/util: Add new cache types to cache entryid -> docid (and reverse)
We define 2 new talloc based types here to allow clients to cache in
memory partial data (just the key/values they are using) held in a global tdb.
This gives us is the ability to have many clients point to cached
values owned/stored in the global tdb, when all clients pointing to a
specific key/value are freed then the forward and reverse mappings
in the global tdb have their reference counts decremented and depending
on the value of the reference count maybe are also removed.

Signed-off-by: Noel Power <noel.power@suse.com>
Reviewed-by: Ralph Boehme <slow@samba.org>